requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-cors-headers==2.4.0
django-environ==0.4.5
django-geoposition==0.3.0
django-google-maps==0.10.1
django-heroku==0.3.1
django-leaflet==0.24.0
django-map-widgets==0.1.9
django-rest-auth==0.9.3
djangorestframework==3.7.7
djangorestframework-gis==0.13
djangorestframework-jwt==1.11.0
future==0.16.0
GDAL==2.2.4
geocoder==1.38.1
geographiclib==1.49
geopy==1.16.0
gunicorn==19.9.0
idna==2.5
lxml==3.8.0
mysql-connector-python==8.0.11
nltk==3.2.5
oauthlib==2.1.0
olefile==0.44
Pillow==4.2.1
protobuf==3.6.0
psycopg2==2.7.5
pygame==1.9.3
PyJWT==1.6.4
pypptx==0.2.2
python-pptx==0.5.7
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.11
ratelim==0.1.6
requests==2.18.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.21.1
virtualenv==16.0.0
whitenoise==3.3.1
wikipedia==1.4.0
XlsxWriter==0.9.8
